REFACTOR: Game folder structure
