+++ panel/src/game/action/useDeleteGame.js
