+++ panel/src/game/settings/GameSettingsViewEdit.js
@@ -1,10 +1,10 @@
 import { Link } from 'react-router-dom';
-import languages from '../shared/helpers/languages';
-import GameLanguageSelect from './create/GameLanguageSelect';
-import NewGamesOnly from './create/NewGamesOnly';
-import EditorOnly from './edit/EditorOnly';
-import { useGameEditorContext } from './edit/GameEditorContext';
-import PrivatePublicToggle from './edit/PrivatePublicToggle/PrivatePublicToggle';
+import languages from '../../shared/helpers/languages';
+import GameLanguageSelect from './GameLanguageSelect';
+import NewGamesOnly from '../wrapper/NewGamesOnly';
+import EditorOnly from '../wrapper/EditorOnly';
+import { useGameEditorContext } from '../GameEditorContext';
+import PrivatePublicToggle from './PrivatePublicToggle/PrivatePublicToggle';
 
 const GameSettingsViewEdit = () => {
   const { author, lng, isPrivate } = useGameEditorContext();
