+++ panel/src/game/wrapper/EditorOnly.js
@@ -1,4 +1,4 @@
-import { useGameEditorContext } from './GameEditorContext';
+import { useGameEditorContext } from '../GameEditorContext';
 
 const EditorOnly = ({ children, shouldNotBeEditor = false }) => {
   const { isEditable } = useGameEditorContext();
